pkg: destroy: define a destroyer-in-stages interface
Each stages is composed of a list of functions that can potentially be
run on parallel. If any of these functions return an error, the stage
state is a failure and the destroy is aborted. If some functions fail
without returning errors, the stage is retried until all of them are
successful.

Providers implementing this interface just need to define their own
stages. A common `Run` function will take care of running each stage
at a time and just proceeding to the next stage when the current one has
succeeded.